Logout problem solved

+++ user_management/api/views.py
@@ -5,6 +5,7 @@ from rest_framework.authtoken.models import Token
 from rest_framework.permissions import IsAuthenticated
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
+from django.urls import reverse
 
 from ..models import Employee
 from .serializers import (UserRegistrationSerializer,UserLoginSerializer)
@@ -47,12 +48,14 @@ class UserLoginApiView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 class UserLogoutView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        request.user.auth_token.delete()
-        logout(request)
-        return Response({'success': "Logout successful"}, status=status.HTTP_200_OK)
+    def get(self, request):
+        if hasattr(request.user, 'auth_token'):
+            request.user.auth_token.delete()
+        login_url = reverse('login')  # This will reverse the 'login' URL name
+        response = Response({'success': "Logout successful"}, status=status.HTTP_200_OK)
+        response['Location'] = login_url  # The Location header now points to the login URL
+        response.status_code = 302  # HTTP Status code for redirection
+        return response
 
 
 class UserListView(APIView):
